requirements.txt
Django==4.2
Pillow==9.5.0
charset-normalizer==3.1.0
django-cors-headers==4.0.0
djangorestframework==3.14.0
requests==2.30.0
sqlparse==0.4.3
reportlab==4.0.4
